PP-59 - CORS fix
