+++ src/main/java/de/htw/gateway/RoutingConfig.java
@@ -1,0 +1,20 @@
+package de.htw.gateway;
+
+import org.springframework.cloud.gateway.route.RouteLocator;
+import org.springframework.cloud.gateway.route.builder.RouteLocatorBuilder;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class RoutingConfig {
+
+    @Bean
+    public RouteLocator customRouteLocator(RouteLocatorBuilder builder) {
+        return builder.routes()
+                .route("product-service", r -> r
+                        .path("/products/**")
+                        .filters(f -> f.prefixPath("/v1"))
+                        .uri("http://product-service:8080"))
+                .build();
+    }
+}
